+++ src/components/UserCollectForm.tsx
@@ -1,34 +1,14 @@
 import { Link } from 'react-router-dom';
-import React, { useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import {
-  Badge,
   Box,
-  Button,
-  Input,
-  NumberInput,
-  NumberInputField,
-  NumberInputStepper,
-  NumberIncrementStepper,
-  NumberDecrementStepper,
-  Select,
-  Card,
-  CardBody,
-  CardFooter,
-  SimpleGrid,
-  Stat,
-  StatLabel,
-  StatNumber,
-  StatHelpText,
-  Table,
-  Thead,
-  Tbody,
-  Heading,
-  Tr,
-  Th,
-  Td,
-  TableContainer,
-  Text,
 } from '@chakra-ui/react';
+
+import ZipCodeStep from './form-steps/ZipCodeStep';
+import CollectPointStep from './form-steps/CollectPointStep';
+import GarbageBagStep from './form-steps/GarbageBagStep';
+import CheckoutStep from './form-steps/CheckoutStep';
+import ConfirmationStep from './form-steps/ConfirmationStep';
 
 export default function UserCollectForm({
   step,
@@ -91,7 +71,7 @@ export default function UserCollectForm({
     },
   ]);
 
-  const [selectedCollectPoint, setselectedCollectPoint] = useState({
+  const [selectedCollectPoint, setSelectedCollectPoint] = useState({
     id: 0,
     name: '',
     address: '',
@@ -117,7 +97,7 @@ export default function UserCollectForm({
   const handlePoints = () => {
     const clearAllState = () => {
       setCEP('');
-      setselectedCollectPoint({
+      setSelectedCollectPoint({
         id: 0,
         name: '',
         address: '',
@@ -137,10 +117,6 @@ export default function UserCollectForm({
     };
     clearAllState();
     setPoints((points as any) + totalWeight);
-  };
-
-  const handleSelectCollectPoint = (collectPoint: any) => {
-    setselectedCollectPoint(collectPoint);
   };
 
   const handleGarbageBags = (action: string) => {
@@ -208,245 +184,19 @@ export default function UserCollectForm({
   return (
     <Box w="360px" h="100%" m="0 auto">
       {step === 'setCEP' && (
-        <Card w="100%" h="100%" p="4">
-          <CardBody>
-            <Heading as="h1" size="lg" mt="4">
-              informe seu CEP
-            </Heading>
-            <Text mt="4">para encontrar pontos de coleta próximos a você</Text>
-            <Input
-              aria-label="CEP"
-              type="number"
-              value={CEP}
-              onChange={(e) => setCEP(e.target.value)}
-              mt="4"
-            />
-            <Link to="/activity/reciclagem/selecao">
-              <Button
-                colorScheme="orange"
-                variant="solid"
-                disabled={!CEP || CEP.length < 8}
-                p="10px 30px"
-                m="20px 0"
-              >
-                próxima etapa
-              </Button>
-            </Link>
-          </CardBody>
-        </Card>
+        <ZipCodeStep CEP={CEP} setCEP={setCEP}/>
       )}
       {step === 'selecao' && (
-        <div>
-          <Heading as="h1" size="lg" mt="4">
-            posto de coleta
-          </Heading>
-          <Text mt="4">selecione o posto de coleta mais próxima etapa</Text>
-          <SimpleGrid
-            spacing={4}
-            templateColumns="repeat(auto-fill, minmax(200px, 1fr))"
-          >
-            {collectPoints.map((collectPoint) => (
-              <Card
-                mt="5"
-                key={collectPoint.id}
-                onClick={() => handleSelectCollectPoint(collectPoint)}
-                bg={
-                  selectedCollectPoint.id === collectPoint.id
-                    ? 'orange.200'
-                    : 'white'
-                }
-                variant={
-                  selectedCollectPoint.id === collectPoint.id
-                    ? 'filled'
-                    : 'outline'
-                }
-              >
-                <CardBody>
-                  <h2>{collectPoint.name}</h2>
-                  <p>{collectPoint.address}</p>
-                </CardBody>
-                <CardFooter>
-                  <Button
-                    colorScheme="orange"
-                    variant="solid"
-                    onClick={() => handleSelectCollectPoint(collectPoint)}
-                    disabled={selectedCollectPoint.id === collectPoint.id}
-                  >
-                    selecionar
-                  </Button>
-                </CardFooter>
-              </Card>
-            ))}
-          </SimpleGrid>
-          <Link to="/activity/reciclagem/formulario">
-            <Button colorScheme="orange" variant="solid">
-              próxima etapa
-            </Button>
-          </Link>
-        </div>
+        <CollectPointStep collectPoints={collectPoints} setSelectedCollectPoint={setSelectedCollectPoint} selectedCollectPoint={selectedCollectPoint} />
       )}
       {step === 'formulario' && (
-        <Box p={4}>
-          <Box>
-            <Heading as="h1" size="lg" mt="4">
-              monte sua coleta
-            </Heading>
-            <Text mt="4">
-              selecione os tipos de resíduos e suas respectivas quantidades
-            </Text>
-          </Box>
-          <Card w="100%" h="100%" p="4" mt="4">
-            <Stat>
-              <StatLabel>{selectedCollectPoint.name}</StatLabel>
-              <StatNumber>{selectedCollectPoint.city}</StatNumber>
-              <StatHelpText>{selectedCollectPoint.address}</StatHelpText>
-            </Stat>
-          </Card>
-          <Box mt={4}>
-            <Heading as="h2" size="md" mt="4">
-              sacos de resíduos
-            </Heading>
-            <NumberInput value={bagCounter} min={1} max={10}>
-              <NumberInputField />
-              <NumberInputStepper>
-                <NumberIncrementStepper
-                  onClick={() => handleGarbageBags('add')}
-                />
-                <NumberDecrementStepper
-                  onClick={() => handleGarbageBags('remove')}
-                />
-              </NumberInputStepper>
-            </NumberInput>
-          </Box>
-          <div>
-          </div>
-          <SimpleGrid
-            mt={6}
-            spacing={4}
-            templateColumns="repeat(auto-fill, minmax(200px, 1fr))"
-          >
-            {garbageBags.map((garbageBag) => (
-              <Card
-                key={garbageBag.id}
-                px={4}
-                py={4}
-                overflow="hidden"
-                variant="outline"
-              >
-                <h4>Saco</h4>
-                <Badge
-                  variant="solid"
-                  colorScheme={handleGarbageBagTypeColor(garbageBag.bagType)}
-                >
-                  {garbageBag.bagType || 'Tipo'}
-                </Badge>
-                <div>
-                  <label htmlFor="bagType">Tipo</label>
-                  <Select
-                    title="tipo"
-                    name="bagType"
-                    id="bagType"
-                    value={garbageBag.bagType}
-                    onChange={(e) => handleGarbageBagType(e, garbageBag.id)}
-                    required
-                  >
-                    <option value="">selecione</option>
-                    <option value="papel">Papel</option>
-                    <option value="plastico">Plástico</option>
-                    <option value="vidro">Vidro</option>
-                    <option value="metal">Metal</option>
-                    <option value="organico">Orgânico</option>
-                  </Select>
-                </div>
-                <div>
-                  <label htmlFor="bagWeight">Peso</label>
-                  <NumberInput
-                    value={garbageBag.bagWeight}
-                    min={0.1}
-                    max={10}
-                    step={0.1}
-                    onChange={(e) => {
-                      handleGarbageBagWeight(e, garbageBag.id);
-                    }}
-                  >
-                    <NumberInputField />
-                    <NumberInputStepper>
-                      <NumberIncrementStepper />
-                      <NumberDecrementStepper />
-                    </NumberInputStepper>
-                  </NumberInput>
-                  <span>kgs</span>
-                </div>
-              </Card>
-            ))}
-          </SimpleGrid>
-          <Link to="/activity/reciclagem/checkout">
-            <Button colorScheme="orange" variant="solid" disabled={!isValid}>
-              próxima etapa
-            </Button>
-          </Link>
-        </Box>
+        <GarbageBagStep garbageBags={garbageBags} handleGarbageBags={handleGarbageBags} handleGarbageBagType={handleGarbageBagType} handleGarbageBagWeight={handleGarbageBagWeight} handleGarbageBagTypeColor={handleGarbageBagTypeColor} isValid={isValid} selectedCollectPoint={selectedCollectPoint} bagCounter={bagCounter} />
       )}
       {step === 'checkout' && (
-        <Box p={4}>
-          <Heading as="h1" size="lg" mt="4">
-            checkout
-          </Heading>
-          <Text>confira os dados da sua coleta</Text>
-          <Card w="100%" h="100%" p="4" bg="orange.500" color="white" mt={4}>
-            <Stat>
-              <StatLabel>{selectedCollectPoint.name}</StatLabel>
-              <StatNumber>{selectedCollectPoint.city}</StatNumber>
-              <StatHelpText>{selectedCollectPoint.address}</StatHelpText>
-            </Stat>
-          </Card>
-          <Box mt={4}>
-            <TableContainer>
-              <Table>
-                <Thead>
-                  <Tr>
-                    <Th>Tipo</Th>
-                    <Th>Peso</Th>
-                  </Tr>
-                </Thead>
-                <Tbody>
-                  {garbageBags.map((garbageBag) => (
-                    <Tr key={garbageBag.id}>
-                      <Td>{garbageBag.bagType}</Td>
-                      <Td>{garbageBag.bagWeight}</Td>
-                    </Tr>
-                  ))}
-                </Tbody>
-              </Table>
-            </TableContainer>
-            <Link to="/activity/reciclagem/confirmacao">
-              <Button colorScheme="orange" variant="solid">
-                próxima etapa
-              </Button>
-            </Link>
-          </Box>
-        </Box>
+       <CheckoutStep garbageBags={garbageBags} selectedCollectPoint={selectedCollectPoint} />
       )}
       {step === 'confirmacao' && (
-        <Box p={4}>
-          <Heading as="h1" size="lg" mt="4">
-            confirmação / resgatar pontos
-          </Heading>
-          <Text mt={4}>
-            agora é só aguardar a confirmação do ponto de coleta para receber
-            seus pontos!
-          </Text>
-          <Link to="/">
-            <Button
-              colorScheme="orange"
-              variant="solid"
-              onClick={handlePoints}
-              mt={5}
-            >
-              Confirmar
-            </Button>
-          </Link>
-        </Box>
+        <ConfirmationStep handlePoints={handlePoints} />
       )}
     </Box>
   );
